Add loading progress bar
Large graphs can take a long time to render, so it's nice to have some
kind of feedback.

It would be possible to reduce the amount of off-screen graph
stabilisation, and make the graph appear earlier before it's closer to a
final shape, but it's quicker to do the calculation off-screen.