Add tooltip on node hover
This tooltip lists the group name, which might help to find the node in
the editor (especially in cases where the colours might be confused).

The tags are also displayed directly. These are something of an
implementation detail at the moment, but it can't hurt to display them.

The fact that I've had to put HTML code into this state object to
achieve this suggests that I'll need to think more about refactoring
this module in future.